requirements.txt
accelerate==1.2.1
huggingface_hub==0.27.0
python-dotenv==1.0.1
omegaconf==2.3.0
numpy==2.2.0
pandas==2.2.3
scikit-learn==1.6.0
tabulate==0.9.0
transformers==4.47.0
torch==2.6.0
tqdm==4.67.1
wandb==0.17.8
keybert==0.8.5
h5py==3.12.1
matplotlib==3.10.0
plotly==5.24.1
seaborn==0.13.2
timm==0.4.5
